Added userInput.cs
getKeyPress() will return a character currently being pressed on the keyboard as a char. mouseClick() will return a Point type set to (-100, -100) until the left mouse button is pressed and released, it will then return a Point type set to the current x and y possition of the mouse cursor on the screen.